filterscopes: create a filterscopes pkg (#2738)
A filterscopes pkg is needed so that the it can be used outside of the
ebpf pkg. This is needed for the upcoming refactors such as the one
that will allow the filterscopes to be used by the "load objects" and
"symbols collision" events.
